Updated ScriptGenie
Made it so when you pick CraftResource you get the color of that resource on your armor piece by default, you can then change it with the hue control.

Added armor resistances to the exported script. Users can now add resistances to their armor.

Updated the UI so its not so boring...